requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
paho-mqtt==1.6.1
onnxruntime==1.15.1
numpy==1.26.0